Added improved acceptance tests :sparkles:
