+++ Tests/Acceptance/ID3TagEditorAcceptanceTest.swift
@@ -1,5 +1,5 @@
 //
-//  ID3TagEditorTestAcceptanceTest.swift
+//  ID3TagEditorAcceptanceTest.swift
 //
 //  Created by Fabrizio Duroni on 23/02/2018.
 //  2018 Fabrizio Duroni.
@@ -8,7 +8,7 @@
 import XCTest
 @testable import ID3TagEditor
 
-class ID3TagEditorTestAcceptanceTest: XCTestCase {
+class ID3TagEditorAcceptanceTest: XCTestCase {
     let id3TagEditor = ID3TagEditor()
     
     func testFailWrongFilePathFilePath() {
@@ -179,6 +179,123 @@ class ID3TagEditorTestAcceptanceTest: XCTestCase {
         XCTAssertEqual((id3Tag?.frames[.Album] as? ID3FrameWithStringContent)?.content, "Vision Songs Vol. 1")
         XCTAssertEqual((id3Tag?.frames[.RecordingYear] as? ID3FrameRecordingYear)?.year, 2018)
         XCTAssertEqual((id3Tag?.frames[.TrackPosition] as? ID3FramePartOfTotal)?.part, 10)
+    }
+    
+    func testReadNewFramesV2() {
+        let id3Tag = try! id3TagEditor.read(from: PathLoader().pathFor(name: "example-newframes-v2-written", fileType: "mp3"))
+        
+        XCTAssertEqual(id3Tag?.properties.version, .version2)
+        XCTAssertEqual(id3Tag?.frames[.Composer]?.id3Identifier, "TCM")
+        XCTAssertEqual((id3Tag?.frames[.Composer] as? ID3FrameWithStringContent)?.content, "Composer V2")
+        XCTAssertEqual(id3Tag?.frames[.Conductor]?.id3Identifier, "TP3")
+        XCTAssertEqual((id3Tag?.frames[.Conductor] as? ID3FrameWithStringContent)?.content, "Conductor V2")
+        XCTAssertEqual(id3Tag?.frames[.ContentGrouping]?.id3Identifier, "TT1")
+        XCTAssertEqual((id3Tag?.frames[.ContentGrouping] as? ID3FrameWithStringContent)?.content, "ContentGrouping V2")
+        XCTAssertEqual(id3Tag?.frames[.Copyright]?.id3Identifier, "TCR")
+        XCTAssertEqual((id3Tag?.frames[.Copyright] as? ID3FrameWithStringContent)?.content, "Copyright V2")
+        XCTAssertEqual(id3Tag?.frames[.EncodedBy]?.id3Identifier, "TEN")
+        XCTAssertEqual((id3Tag?.frames[.EncodedBy] as? ID3FrameWithStringContent)?.content, "EncodedBy V2")
+        XCTAssertEqual(id3Tag?.frames[.EncoderSettings]?.id3Identifier, "TSS")
+        XCTAssertEqual((id3Tag?.frames[.EncoderSettings] as? ID3FrameWithStringContent)?.content, "EncoderSettings V2")
+        XCTAssertEqual(id3Tag?.frames[.Lyricist]?.id3Identifier, "TXT")
+        XCTAssertEqual((id3Tag?.frames[.Lyricist] as? ID3FrameWithStringContent)?.content, "Lyricist V2")
+        XCTAssertEqual(id3Tag?.frames[.MixArtist]?.id3Identifier, "TP4")
+        XCTAssertEqual((id3Tag?.frames[.MixArtist] as? ID3FrameWithStringContent)?.content, "MixArtist V2")
+        XCTAssertEqual(id3Tag?.frames[.Publisher]?.id3Identifier, "TPB")
+        XCTAssertEqual((id3Tag?.frames[.Publisher] as? ID3FrameWithStringContent)?.content, "Publisher V2")
+        XCTAssertEqual(id3Tag?.frames[.Subtitle]?.id3Identifier, "TT3")
+        XCTAssertEqual((id3Tag?.frames[.Subtitle] as? ID3FrameWithStringContent)?.content, "Subtitle V2")
+        XCTAssertEqual(id3Tag?.frames[.DiscPosition]?.id3Identifier, "TPA")
+        XCTAssertEqual((id3Tag?.frames[.DiscPosition] as? ID3FramePartOfTotal)?.part, 1)
+        XCTAssertEqual((id3Tag?.frames[.DiscPosition] as? ID3FramePartOfTotal)?.total, 3)
+    }
+    
+    func testReadNewFramesV3() {
+        let id3Tag = try! id3TagEditor.read(from: PathLoader().pathFor(name: "example-newframes-v3-written", fileType: "mp3"))
+        
+        XCTAssertEqual(id3Tag?.properties.version, .version3)
+        XCTAssertEqual(id3Tag?.frames[.Composer]?.id3Identifier, "TCOM")
+        XCTAssertEqual((id3Tag?.frames[.Composer] as? ID3FrameWithStringContent)?.content, "Composer V3")
+        XCTAssertEqual(id3Tag?.frames[.Conductor]?.id3Identifier, "TPE3")
+        XCTAssertEqual((id3Tag?.frames[.Conductor] as? ID3FrameWithStringContent)?.content, "Conductor V3")
+        XCTAssertEqual(id3Tag?.frames[.Copyright]?.id3Identifier, "TCOP")
+        XCTAssertEqual((id3Tag?.frames[.Copyright] as? ID3FrameWithStringContent)?.content, "Copyright V3")
+        XCTAssertEqual(id3Tag?.frames[.ContentGrouping]?.id3Identifier, "TIT1")
+        XCTAssertEqual((id3Tag?.frames[.ContentGrouping] as? ID3FrameWithStringContent)?.content, "ContentGrouping V3")
+        XCTAssertEqual(id3Tag?.frames[.DiscPosition]?.id3Identifier, "TPOS")
+        XCTAssertEqual((id3Tag?.frames[.DiscPosition] as? ID3FramePartOfTotal)?.part, 1)
+        XCTAssertEqual((id3Tag?.frames[.DiscPosition] as? ID3FramePartOfTotal)?.total, 3)
+        XCTAssertEqual(id3Tag?.frames[.EncodedBy]?.id3Identifier, "TENC")
+        XCTAssertEqual((id3Tag?.frames[.EncodedBy] as? ID3FrameWithStringContent)?.content, "EncodedBy V3")
+        XCTAssertEqual(id3Tag?.frames[.EncoderSettings]?.id3Identifier, "TSSE")
+        XCTAssertEqual((id3Tag?.frames[.EncoderSettings] as? ID3FrameWithStringContent)?.content, "EncoderSettings V3")
+        XCTAssertEqual(id3Tag?.frames[.iTunesGrouping]?.id3Identifier, "GRP1")
+        XCTAssertEqual((id3Tag?.frames[.iTunesGrouping] as? ID3FrameWithStringContent)?.content, "ItunesGrouping V3")
+        XCTAssertEqual(id3Tag?.frames[.Lyricist]?.id3Identifier, "TEXT")
+        XCTAssertEqual((id3Tag?.frames[.Lyricist] as? ID3FrameWithStringContent)?.content, "Lyricist V3")
+        XCTAssertEqual(id3Tag?.frames[.MixArtist]?.id3Identifier, "TPE4")
+        XCTAssertEqual((id3Tag?.frames[.MixArtist] as? ID3FrameWithStringContent)?.content, "MixArtist V3")
+        XCTAssertEqual(id3Tag?.frames[.FileOwner]?.id3Identifier, "TOWN")
+        XCTAssertEqual((id3Tag?.frames[.FileOwner] as? ID3FrameWithStringContent)?.content, "FileOwner V3")
+        XCTAssertEqual(id3Tag?.frames[.iTunesMovementName]?.id3Identifier, "MVNM")
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementName] as? ID3FrameWithStringContent)?.content, "MovementName V3")
+        XCTAssertEqual(id3Tag?.frames[.iTunesMovementIndex]?.id3Identifier, "MVIN")
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementIndex] as? ID3FrameWithIntegerContent)?.value, 6)
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementCount] as? ID3FrameWithIntegerContent)?.value, 13)
+        XCTAssertEqual(id3Tag?.frames[.iTunesPodcastCategory]?.id3Identifier, "TCAT")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastCategory] as? ID3FrameWithStringContent)?.content, "PodcastCategory V3")
+        XCTAssertEqual(id3Tag?.frames[.iTunesPodcastDescription]?.id3Identifier, "TDES")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastDescription] as? ID3FrameWithStringContent)?.content, "PodcastDescription V3")
+        XCTAssertEqual(id3Tag?.frames[.iTunesPodcastID]?.id3Identifier, "TGID")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastID] as? ID3FrameWithStringContent)?.content, "PodcastID V3")
+        XCTAssertEqual(id3Tag?.frames[.iTunesPodcastKeywords]?.id3Identifier, "TKWD")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastKeywords] as? ID3FrameWithStringContent)?.content, "PodcastKeywords V3")
+        XCTAssertEqual(id3Tag?.frames[.Publisher]?.id3Identifier, "TPUB")
+        XCTAssertEqual((id3Tag?.frames[.Publisher] as? ID3FrameWithStringContent)?.content, "Publisher V3")
+        XCTAssertEqual(id3Tag?.frames[.Subtitle]?.id3Identifier, "TIT3")
+        XCTAssertEqual((id3Tag?.frames[.Subtitle] as? ID3FrameWithStringContent)?.content, "Subtitle V3")
+    }
+    
+    func testReadNewFramesV4() {
+        let id3Tag = try! id3TagEditor.read(from: PathLoader().pathFor(name: "example-newframes-v4-written", fileType: "mp3"))
+        
+        XCTAssertEqual(id3Tag?.properties.version, .version4)
+        XCTAssertEqual((id3Tag?.frames[.Composer] as? ID3FrameWithStringContent)?.id3Identifier, "TCOM")
+        XCTAssertEqual((id3Tag?.frames[.Composer] as? ID3FrameWithStringContent)?.content, "Composer V4")
+        XCTAssertEqual((id3Tag?.frames[.Conductor] as? ID3FrameWithStringContent)?.id3Identifier, "TPE3")
+        XCTAssertEqual((id3Tag?.frames[.Conductor] as? ID3FrameWithStringContent)?.content, "Conductor V4")
+        XCTAssertEqual((id3Tag?.frames[.Copyright] as? ID3FrameWithStringContent)?.id3Identifier, "TCOP")
+        XCTAssertEqual((id3Tag?.frames[.Copyright] as? ID3FrameWithStringContent)?.content, "Copyright V4")
+        XCTAssertEqual((id3Tag?.frames[.ContentGrouping] as? ID3FrameWithStringContent)?.id3Identifier, "TIT1")
+        XCTAssertEqual((id3Tag?.frames[.ContentGrouping] as? ID3FrameWithStringContent)?.content, "ContentGrouping V4")
+        XCTAssertEqual((id3Tag?.frames[.EncodedBy] as? ID3FrameWithStringContent)?.id3Identifier, "TENC")
+        XCTAssertEqual((id3Tag?.frames[.EncodedBy] as? ID3FrameWithStringContent)?.content, "EncodedBy V4")
+        XCTAssertEqual((id3Tag?.frames[.EncoderSettings] as? ID3FrameWithStringContent)?.id3Identifier, "TSSE")
+        XCTAssertEqual((id3Tag?.frames[.EncoderSettings] as? ID3FrameWithStringContent)?.content, "EncoderSettings V4")
+        XCTAssertEqual((id3Tag?.frames[.FileOwner] as? ID3FrameWithStringContent)?.id3Identifier, "TOWN")
+        XCTAssertEqual((id3Tag?.frames[.FileOwner] as? ID3FrameWithStringContent)?.content, "FileOwner V4")
+        XCTAssertEqual((id3Tag?.frames[.iTunesGrouping] as? ID3FrameWithStringContent)?.id3Identifier, "GRP1")
+        XCTAssertEqual((id3Tag?.frames[.iTunesGrouping] as? ID3FrameWithStringContent)?.content, "ItunesGrouping V4")
+        XCTAssertEqual((id3Tag?.frames[.Lyricist] as? ID3FrameWithStringContent)?.id3Identifier, "TEXT")
+        XCTAssertEqual((id3Tag?.frames[.Lyricist] as? ID3FrameWithStringContent)?.content, "Lyricist V4")
+        XCTAssertEqual((id3Tag?.frames[.MixArtist] as? ID3FrameWithStringContent)?.id3Identifier, "TPE4")
+        XCTAssertEqual((id3Tag?.frames[.MixArtist] as? ID3FrameWithStringContent)?.content, "MixArtist V4")
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementName] as? ID3FrameWithStringContent)?.id3Identifier, "MVNM")
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementName] as? ID3FrameWithStringContent)?.content, "MovementName V4")
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementIndex] as? ID3FrameWithIntegerContent)?.value, 6)
+        XCTAssertEqual((id3Tag?.frames[.iTunesMovementCount] as? ID3FrameWithIntegerContent)?.value, 13)
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastCategory] as? ID3FrameWithStringContent)?.id3Identifier, "TCAT")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastCategory] as? ID3FrameWithStringContent)?.content, "PodcastCategory V4")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastDescription] as? ID3FrameWithStringContent)?.id3Identifier, "TDES")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastDescription] as? ID3FrameWithStringContent)?.content, "PodcastDescription V4")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastID] as? ID3FrameWithStringContent)?.id3Identifier, "TGID")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastID] as? ID3FrameWithStringContent)?.content, "PodcastID V4")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastKeywords] as? ID3FrameWithStringContent)?.id3Identifier, "TKWD")
+        XCTAssertEqual((id3Tag?.frames[.iTunesPodcastKeywords] as? ID3FrameWithStringContent)?.content, "PodcastKeywords V4")
+        XCTAssertEqual((id3Tag?.frames[.Publisher] as? ID3FrameWithStringContent)?.id3Identifier, "TPUB")
+        XCTAssertEqual((id3Tag?.frames[.Publisher] as? ID3FrameWithStringContent)?.content, "Publisher V4")
+        XCTAssertEqual((id3Tag?.frames[.Subtitle] as? ID3FrameWithStringContent)?.id3Identifier, "TIT3")
+        XCTAssertEqual((id3Tag?.frames[.Subtitle] as? ID3FrameWithStringContent)?.content, "Subtitle V4")
     }
     
     
@@ -459,6 +576,141 @@ class ID3TagEditorTestAcceptanceTest: XCTestCase {
             try! Data(contentsOf: URL(fileURLWithPath: pathMp3Generated)),
             try! Data(contentsOf: URL(fileURLWithPath: pathMp3ToCompare))
         )
+    }
+    
+    func testWriteNewFramesV2() {
+        let pathMp3ToCompare = PathLoader().pathFor(name: "example-newframes-v2-written", fileType: "mp3")
+        let pathMp3Generated = NSHomeDirectory() + "/example-newframes-v2-generated.mp3"
+        let id3Tag = ID3Tag(
+            version: .version2,
+            frames: [
+                .Composer : ID3FrameWithStringContent(content: "Composer V2"),
+                .Conductor : ID3FrameWithStringContent(content: "Conductor V2"),
+                .ContentGrouping : ID3FrameWithStringContent(content: "ContentGrouping V2"),
+                .Copyright : ID3FrameWithStringContent(content: "Copyright V2"),
+                .DiscPosition : ID3FramePartOfTotal(part: 1, total: 3),
+                .EncodedBy : ID3FrameWithStringContent(content: "EncodedBy V2"),
+                .EncoderSettings : ID3FrameWithStringContent(content: "EncoderSettings V2"),
+                .Lyricist : ID3FrameWithStringContent(content: "Lyricist V2"),
+                .MixArtist : ID3FrameWithStringContent(content: "MixArtist V2"),
+                .Publisher : ID3FrameWithStringContent(content: "Publisher V2"),
+                .Subtitle : ID3FrameWithStringContent(content: "Subtitle V2"),
+            ]
+        )
+        
+        XCTAssertNoThrow(try id3TagEditor.write(
+            tag: id3Tag,
+            to: PathLoader().pathFor(name: "example-newframes", fileType: "mp3"),
+            andSaveTo: pathMp3Generated
+        ))
+        XCTAssertEqual(
+            try! Data(contentsOf: URL(fileURLWithPath: pathMp3Generated)),
+            try! Data(contentsOf: URL(fileURLWithPath: pathMp3ToCompare))
+        )
+    }
+    
+    func testWriteNewFramesV3() {
+        let pathMp3ToCompare = PathLoader().pathFor(name: "example-newframes-v3-written", fileType: "mp3")
+        let pathMp3Generated = NSHomeDirectory() + "/example-newframes-v3-generated.mp3"
+        let id3Tag = ID3Tag(
+            version: .version3,
+            frames: [
+                .Composer : ID3FrameWithStringContent(content: "Composer V3"),
+                .Conductor : ID3FrameWithStringContent(content: "Conductor V3"),
+                .ContentGrouping : ID3FrameWithStringContent(content: "ContentGrouping V3"),
+                .Copyright : ID3FrameWithStringContent(content: "Copyright V3"),
+                .DiscPosition : ID3FramePartOfTotal(part: 1, total: 3),
+                .EncodedBy : ID3FrameWithStringContent(content: "EncodedBy V3"),
+                .EncoderSettings : ID3FrameWithStringContent(content: "EncoderSettings V3"),
+                .FileOwner : ID3FrameWithStringContent(content: "FileOwner V3"),
+                .Lyricist : ID3FrameWithStringContent(content: "Lyricist V3"),
+                .MixArtist : ID3FrameWithStringContent(content: "MixArtist V3"),
+                .Subtitle : ID3FrameWithStringContent(content: "Subtitle V3"),
+                .Publisher : ID3FrameWithStringContent(content: "Publisher V3"),
+                .iTunesGrouping : ID3FrameWithStringContent(content: "ItunesGrouping V3"),
+                .iTunesMovementName : ID3FrameWithStringContent(content: "MovementName V3"),
+                .iTunesMovementIndex : ID3FrameWithIntegerContent(value: 6),
+                .iTunesMovementCount : ID3FrameWithIntegerContent(value: 13),
+                .iTunesPodcastCategory : ID3FrameWithStringContent(content: "PodcastCategory V3"),
+                .iTunesPodcastDescription : ID3FrameWithStringContent(content: "PodcastDescription V3"),
+                .iTunesPodcastID : ID3FrameWithStringContent(content: "PodcastID V3"),
+                .iTunesPodcastKeywords : ID3FrameWithStringContent(content: "PodcastKeywords V3"),
+            ]
+        )
+        
+        XCTAssertNoThrow(try id3TagEditor.write(
+            tag: id3Tag,
+            to: PathLoader().pathFor(name: "example-newframes", fileType: "mp3"),
+            andSaveTo: pathMp3Generated
+        ))
+        XCTAssertEqual(
+            try! Data(contentsOf: URL(fileURLWithPath: pathMp3Generated)),
+            try! Data(contentsOf: URL(fileURLWithPath: pathMp3ToCompare))
+        )
+    }
+    
+    func testWriteNewFramesV4() {
+        let pathMp3ToCompare = PathLoader().pathFor(name: "example-newframes-v4-written", fileType: "mp3")
+        let pathMp3Generated = NSHomeDirectory() + "/example-newframes-v4-generated.mp3"
+        let id3Tag = ID3Tag(
+            version: .version4,
+            frames: [
+                .Composer : ID3FrameWithStringContent(content: "Composer V4"),
+                .Conductor : ID3FrameWithStringContent(content: "Conductor V4"),
+                .ContentGrouping : ID3FrameWithStringContent(content: "ContentGrouping V4"),
+                .Copyright : ID3FrameWithStringContent(content: "Copyright V4"),
+                .DiscPosition : ID3FramePartOfTotal(part: 1, total: 3),
+                .EncodedBy : ID3FrameWithStringContent(content: "EncodedBy V4"),
+                .EncoderSettings : ID3FrameWithStringContent(content: "EncoderSettings V4"),
+                .FileOwner : ID3FrameWithStringContent(content: "FileOwner V4"),
+                .Lyricist : ID3FrameWithStringContent(content: "Lyricist V4"),
+                .MixArtist : ID3FrameWithStringContent(content: "MixArtist V4"),
+                .Publisher : ID3FrameWithStringContent(content: "Publisher V4"),
+                .Subtitle : ID3FrameWithStringContent(content: "Subtitle V4"),
+                .iTunesGrouping : ID3FrameWithStringContent(content: "ItunesGrouping V4"),
+                .iTunesMovementName : ID3FrameWithStringContent(content: "MovementName V4"),
+                .iTunesMovementIndex : ID3FrameWithIntegerContent(value: 6),
+                .iTunesMovementCount : ID3FrameWithIntegerContent(value: 13),
+                .iTunesPodcastCategory : ID3FrameWithStringContent(content: "PodcastCategory V4"),
+                .iTunesPodcastDescription : ID3FrameWithStringContent(content: "PodcastDescription V4"),
+                .iTunesPodcastID : ID3FrameWithStringContent(content: "PodcastID V4"),
+                .iTunesPodcastKeywords : ID3FrameWithStringContent(content: "PodcastKeywords V4"),
+            ]
+        )
+        
+        XCTAssertNoThrow(try id3TagEditor.write(
+            tag: id3Tag,
+            to: PathLoader().pathFor(name: "example-newframes", fileType: "mp3"),
+            andSaveTo: pathMp3Generated
+        ))
+        
+        XCTAssertEqual(
+            try! Data(contentsOf: URL(fileURLWithPath: pathMp3Generated)),
+            try! Data(contentsOf: URL(fileURLWithPath: pathMp3ToCompare))
+        )
+    }
+    
+    func testFramesAfterAttachdPicturesAreWritten() {
+        let path = PathLoader().pathFor(name: "folder", fileType: "jpg")
+        let pathMp3Generated = NSHomeDirectory() + "/frames-after-attached-picture.mp3"
+        let cover = try! Data(contentsOf: URL(fileURLWithPath: path))
+        let id3Tag = ID3Tag(
+            version: .version3,
+            frames: [
+                .DiscPosition : ID3FramePartOfTotal(part: 1, total: 3),
+                .AttachedPicture(.FrontCover) : ID3FrameAttachedPicture(picture: cover, type: .FrontCover, format: .Jpeg),
+            ]
+        )
+        
+        let id3TagEditor = ID3TagEditor()
+        
+        XCTAssertNoThrow(try! id3TagEditor.write(tag: id3Tag,
+                                                 to: PathLoader().pathFor(name: "frames-after-attached-picture", fileType: "mp3"),
+                                                 andSaveTo: pathMp3Generated))
+        
+        let tag = try! id3TagEditor.read(from: pathMp3Generated)
+        XCTAssertEqual((tag?.frames[.DiscPosition] as? ID3FramePartOfTotal)?.part, 1)
+        XCTAssertEqual((tag?.frames[.DiscPosition] as? ID3FramePartOfTotal)?.total, 3)
     }
     
     func testWriteTagV4SynchsafeIntegers() {
